Move indextank_client into lib/thinkingtank so it can be required more easily
