+++ lib/thinkingtank/indextank_client.rb
